moved ball class to nother file, some bugfixes
