+++ cuestick.py
@@ -4,11 +4,11 @@ import physics
 
 
 def set_cue(gameState,ball_id):
-    hit_power = 1/3.0
+    hit_power = 0.8
     cue_length = 300
     cue_thickness = 3
 
-    def draw_cue(ball_id, cue_angle, displacement, color):
+    def draw_cue( cue_angle, displacement, color):
         # dy/dx = sin(a)/cos(s) = tan(a)
         cos_a = math.cos(math.radians(cue_angle))
         sin_a = math.sin(math.radians(cue_angle))
@@ -30,8 +30,8 @@ def set_cue(gameState,ball_id):
 
         return points
 
-    def delete_cue(ball_id, angle, displacement):
-        draw_cue(ball_id, angle, displacement,
+    def delete_cue( angle, displacement):
+        draw_cue( angle, displacement,
                  gameState.table_color)
         gameState.canvas.draw_table_sides(gameState)
         gameState.canvas.draw_table_holes(gameState)
@@ -49,8 +49,8 @@ def set_cue(gameState,ball_id):
                 (line_y > gameState.table_margin) and (line_y < gameState.canvas.size_y - gameState.table_margin):
             pygame.draw.line(gameState.canvas.surface, color, (line_x, line_y),
                              (line_x + line_dist * cos_a, line_y + line_dist * sin_a))
-            line_x = line_x + 2 * line_dist * cos_a
-            line_y = line_y + 2 * line_dist * sin_a
+            line_x += 2 * line_dist * cos_a
+            line_y += 2 * line_dist * sin_a
             count += 1
 
     def get_cue_displacement(mouse_pos, ball, initial_distance):
@@ -68,11 +68,10 @@ def set_cue(gameState,ball_id):
     ball = gameState.balls[ball_id]
     angle = 0
     prev_angle = angle
-    rect_pointlist = draw_cue(ball_id, angle, ball.size, (100, 100, 100))
+    rect_pointlist = draw_cue(angle, ball.size, (100, 100, 100))
     pygame.display.update()
 
     start_pos = pygame.mouse.get_pos()
-    final_pos = start_pos
     displacement = 0
 
     done = False
@@ -81,9 +80,10 @@ def set_cue(gameState,ball_id):
         pygame.event.get()
         if pygame.mouse.get_pressed()[0] and physics.is_point_in_rect(rect_pointlist, start_pos):
             done = True
-            final_pos = start_pos
             inital_mouse_dist = physics.point_distance(start_pos, (ball.x, ball.y))
             prev_displacement = ball.size
+            gameState.mark_one_frame()
+
             # cue was displaced from the cue ball
             while pygame.mouse.get_pressed()[0]:
                 pygame.event.get()
@@ -101,28 +101,28 @@ def set_cue(gameState,ball_id):
                 if dx > 0:
                     angle -= 180
                 if not (prev_angle == angle) or not (prev_displacement == displacement):
-                    delete_cue(ball_id, prev_angle, prev_displacement)
+                    delete_cue( prev_angle, prev_displacement)
                     draw_lines(ball, prev_angle + 180, gameState.table_color)
-                    rect_pointlist = draw_cue(ball_id, angle, displacement, gameState.cue_color)
+                    rect_pointlist = draw_cue(angle, displacement, gameState.cue_color)
                     draw_lines(ball, angle + 180, (255, 255, 255))
                     pygame.display.update()
                     prev_angle = angle
                     prev_displacement = displacement
 
+                gameState.mark_one_frame()
 
             if (displacement == ball.size):
                 done = False
 
     draw_lines(ball, prev_angle + 180, gameState.table_color)
-    # small hitting animation
+    # hitting animation
     prev_n = displacement
     for n in range(int(displacement), ball.size, -1):
-        delete_cue(ball_id, angle, prev_n)
-        draw_cue(ball_id, angle, n, gameState.cue_color)
+        delete_cue(angle, prev_n)
+        draw_cue(angle, n, gameState.cue_color)
         pygame.display.update()
         prev_n = n
-        # time.sleep(0.5)
-    delete_cue(ball_id, prev_angle, gameState.ball_size)
+    delete_cue(prev_angle, gameState.ball_size)
 
     sin_a = math.sin(math.radians(angle))
     cos_a = math.cos(math.radians(angle))
